Simplify first setup
Make sure that first setup works even when we are in broken state. Also as we
don't support offline migration anymore, so let's drop the relevant code. The
only migration we do is from the root to the external storage, therefor we
don't need to search for UUID of srv here as init script tries to find it by
itself if old_uuid is rootfs.